% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_utils.R
\name{capture_shortcuts_to_package}
\alias{capture_shortcuts_to_package}
\title{Grab the Rstudio keyboard shortcuts from the local user file structure
and save them to a variable in the package}
\usage{
capture_shortcuts_to_package(shortcuts)
}
\arguments{
\item{overwrite}{}
}
\value{

}
\description{
Grab the Rstudio keyboard shortcuts from the local user file structure
and save them to a variable in the package
}
