% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_utils.R
\name{get_keyboard_shortcuts}
\alias{get_keyboard_shortcuts}
\title{Grab the RStudio keyboard shortcuts and return them as a named list}
\usage{
get_keyboard_shortcuts(local = F)
}
\arguments{
\item{local}{}
}
\value{
list containing three named lists
}
\description{
Grab the RStudio keyboard shortcuts and return them as a named list
}
