% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_os}
\alias{get_os}
\title{Determines the host operating system
TODO: not clear if this works for windows
SOURCE: https://www.r-bloggers.com/identifying-the-os-from-r/}
\usage{
get_os()
}
\value{
one of "osx", "linux"
}
\description{
Determines the host operating system
TODO: not clear if this works for windows
SOURCE: https://www.r-bloggers.com/identifying-the-os-from-r/
}
\examples{

get_os()
}
