% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_utils.R
\name{get_shortcut_dir}
\alias{get_shortcut_dir}
\title{Determine the directory name for RStudio keybindings}
\usage{
get_shortcut_dir()
}
\value{
a string containing the name of the directory containing the keybindings
}
\description{
Determine the directory name for RStudio keybindings
}
\examples{

get_shortcut_dir()

}
