% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_utils.R
\name{get_shortcut_filenames}
\alias{get_shortcut_filenames}
\title{Build a list which contains three named lists, which are the names of the three
files ued for RStudio keyboard shortcuts}
\usage{
get_shortcut_filenames(local = F)
}
\arguments{
\item{local}{}
}
\description{
Build a list which contains three named lists, which are the names of the three
files ued for RStudio keyboard shortcuts
}
\examples{
get_shortcut_filenames()

get_shortcut_filenames(local=T)

}
