% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_utils.R
\name{restore_keyboard_shortcuts}
\alias{restore_keyboard_shortcuts}
\title{restores RStudio keyboard shortcuts from the data saved to the package
in the 'shortcuts' variable}
\usage{
restore_keyboard_shortcuts(shortcuts, local = T, overwrite = F)
}
\arguments{
\item{overwrite}{}
}
\value{

}
\description{
restores RStudio keyboard shortcuts from the data saved to the package
in the 'shortcuts' variable
}
